fix: moved types to a better location
